chore(credential): split credential module
Split the `credential` crate from a monolithic `src/lib.rs` into a core
`src/lib.rs` providing things like the `ProvideAwsCredentials` trait, and
modules providing actual AWS credential provider implementations (e.g.
`profile::ProfileProvider`).
